+++ ProgettoIngegneria/src/main/java/com/example/progettoingegneria/FinestraEsperienzaLavorativa.java
@@ -60,7 +60,7 @@ public class FinestraEsperienzaLavorativa extends Application{
             public void handle(ActionEvent event) {
                 LocalDate inizioPeriodoLavorativo=LocalDate.of(Integer.parseInt(annoInizio.getText()),Integer.parseInt(meseInizio.getText()),Integer.parseInt(giornoInizio.getText()));
                 LocalDate finePeriodoLavorativo=LocalDate.of(Integer.parseInt(annoFine.getText()),Integer.parseInt(meseFine.getText()),Integer.parseInt(giornoFine.getText()));
-                EsperienzaLavorativa esp=EsperienzaLavorativa.of(inizioPeriodoLavorativo,finePeriodoLavorativo,nome.getText(), Arrays.asList(mansioni.getText().split(",")),ubicazione.getText(), Integer.parseInt(retribuzione.getText().replaceAll(" ","")));
+                EsperienzaLavorativa esp=EsperienzaLavorativa.of(inizioPeriodoLavorativo,finePeriodoLavorativo,nome.getText(), Arrays.asList(mansioni.getText().split(",")),ubicazione.getText(), Double.parseDouble(retribuzione.getText().replaceAll(" ","")));
 
                 if(modifica==false)
                     esperienze.add(esp);
